web/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-multipart==0.0.9
numpy>=1.24.0
opencv-python-headless>=4.8.0
insightface>=0.7.3
onnxruntime>=1.16.0
Pillow>=10.0.0
